ITC-2986 Fix nginx config

+++ CONTAINER_SETUP.sh
@@ -217,7 +217,9 @@ if [[ -d "/opt/openitc/nagios/rollout" ]]; then
 fi
 
 echo "Enabling webserver configuration"
-ln -s /etc/nginx/sites-available/openitc /etc/nginx/sites-enabled/openitc
+if [[ ! -f "/etc/nginx/sites-enabled/openitc" ]]; then
+    ln -s /etc/nginx/sites-available/openitc /etc/nginx/sites-enabled/openitc
+fi
 rm -f /etc/nginx/sites-enabled/default
 
 #Set default permissions, check for always allowed permissions and dependencies
